OO-106: separated buchhaltung repositories
